+++ src/anemoi/transform/filters/rename.py
@@ -21,7 +21,12 @@ class FormatRename:
     def __init__(self, what, format):
         self.what = what
         self.format = format
-        self.bits = re.findall(r"{(\w+)}", format)
+        self.bits = re.findall(r"{([\w:]+)}", format)
+
+        # Escape ":" type delimiter used by eccodes as ":" is a reserved symbol in str.format.
+        self._delimiter = "|"
+        self.format = re.sub(r"{([^}]+)}", lambda m: "{" + m.group(1).replace(":", self._delimiter) + "}", self.format)
+        self.format_keys = [b.replace(":", self._delimiter) for b in self.bits]
 
     def rename(self, field):
         md = field.metadata(self.what, default=None)
@@ -37,7 +42,7 @@ class FormatRename:
             else values
         )
 
-        kwargs = {k: v for k, v in zip(self.bits, values)}
+        kwargs = dict(zip(self.format_keys, values))
         kwargs = {self.what: self.format.format(**kwargs)}
         return new_field_with_metadata(template=field, **kwargs)
 
@@ -105,10 +110,12 @@ class Rename(Filter):
                 - source:
                     ...
                 - rename:
-                    param: "{param}_{levelist}_{levtype}"
+                    param: "{param}_{levelist}_{levtype}_{level:d}"
 
     In the latter case, the keys between curly braces are replaced by their
-    corresponding metadata values in the field.
+    corresponding metadata values in the field. The type of metadata values
+    requested via eccodes can be chosen by appending ":i", ":d", ":s" for
+    int, double and str, respectively. (See https://confluence.ecmwf.int/display/ECC/grib_get)
 
     """