+++ CHANGELOG.md
@@ -9,6 +9,32 @@ Please add your functional changes to the appropriate section in the PR.
 Keep it human-readable, your future self will thank you!
 
 
+## [0.1.19](https://github.com/ecmwf/anemoi-transform/compare/0.1.18...0.1.19) (2025-11-07)
+
+
+### Features
+
+* Support formats in rename filter ([#186](https://github.com/ecmwf/anemoi-transform/issues/186)) ([00e9890](https://github.com/ecmwf/anemoi-transform/commit/00e989013018bfde8d8f73da17c0b59265d7760b))
+
+
+### Bug Fixes
+
+* Update README to reflect project maturity status ([#184](https://github.com/ecmwf/anemoi-transform/issues/184)) ([d183f94](https://github.com/ecmwf/anemoi-transform/commit/d183f94c44f3729a97d5d09fe00e865191856973))
+
+## [0.1.18](https://github.com/ecmwf/anemoi-transform/compare/0.1.17...0.1.18) (2025-10-30)
+
+
+### Features
+
+* Add AccumToInterval filter transform ([#168](https://github.com/ecmwf/anemoi-transform/issues/168)) ([0801a4b](https://github.com/ecmwf/anemoi-transform/commit/0801a4b1ea040262efcc4edd654419e0e73ecc54))
+* Generic cos sin from rad filter ([#180](https://github.com/ecmwf/anemoi-transform/issues/180)) ([7fabff9](https://github.com/ecmwf/anemoi-transform/commit/7fabff9d00b685513c36eb4d71aae54368b0d199))
+
+
+### Bug Fixes
+
+* Clarify error message ([#181](https://github.com/ecmwf/anemoi-transform/issues/181)) ([1cd0536](https://github.com/ecmwf/anemoi-transform/commit/1cd0536f68d560cd02644cd9ba305b9db2deebf0))
+* Remove variable from xarray metadata when grouping ([#182](https://github.com/ecmwf/anemoi-transform/issues/182)) ([46465cd](https://github.com/ecmwf/anemoi-transform/commit/46465cdab7518d49e660bacded13763547730d0f))
+
 ## [0.1.17](https://github.com/ecmwf/anemoi-transform/compare/0.1.16...0.1.17) (2025-10-01)